requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pillow
opencv-python
torch
numpy
torchvision